services/graph/requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
neo4j>=5.10.0
structlog>=23.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.20
structlog==23.2.0
prometheus-client==0.19.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Kafka (existing FSMA functionality)
aiokafka==0.8.1
kafka-python==2.0.2

# Utilities
python-dotenv==1.0.0
